File renaming and moving, for aggregate, combine, metrics
