+++ kglib/kgcn/embed/embedding.py
@@ -20,7 +20,8 @@
 import numpy as np
 import tensorflow as tf
 
-import kglib.kgcn.embed.aggregation as agg
+import kglib.kgcn.embed.aggregate as aggregate
+import kglib.kgcn.embed.combine as combine
 
 
 class Embedder:
@@ -37,7 +38,7 @@ class Embedder:
     def __call__(self, neighbourhoods):
 
         # TODO pass through params for aggregators, combiners and normalisers
-        aggregators = [agg.Aggregate(self._aggregated_size, name=f'aggregate_{i}') for i in
+        aggregators = [aggregate.Aggregator(self._aggregated_size, name=f'aggregate_{i}') for i in
                        range(len(self._neighbourhood_sizes))]
 
         combiners = []
@@ -47,19 +48,20 @@ class Embedder:
             #                                     name=f'weights_{i}')
 
             if i + 1 == len(self._neighbourhood_sizes):
-                # combiner = agg.Combine(weights, activation=lambda x: x, name=f'combine_{i}_linear')
-                combiner = agg.DenseCombine(self._embedding_size, activation=lambda x: x, name=f'combine_{i}_linear')
+                # combiner = combine.Combiner(weights, activation=lambda x: x, name=f'combine_{i}_linear')
+                combiner = combine.DenseCombiner(self._embedding_size, activation=lambda x: x, name=f'combine_{i}_linear')
             else:
-                combiner = agg.DenseCombine(self._embedding_size, activation=tf.nn.relu, name=f'combine_{i}_relu')
+                combiner = combine.DenseCombiner(self._embedding_size, activation=tf.nn.relu, name=f'combine_{i}_relu')
             combiners.append(combiner)
 
-        normalisers = [agg.normalise for _ in range(len(self._neighbourhood_sizes))]
+        normalisers = [normalise for _ in range(len(self._neighbourhood_sizes))]
 
-        full_representation = agg.chain_aggregate_combine(neighbourhoods, aggregators, combiners, normalisers)
+        full_representation = chain_aggregate_combine(neighbourhoods, aggregators, combiners, normalisers)
         # full_representation = tf.nn.l2_normalize(full_representation, -1)
         return full_representation
 
 
+# TODO Presently unused
 def initialise_glorot_weights(shape, name=None):
     """
     Glorot & Bengio (AISTATS 2010) init.
@@ -74,3 +76,27 @@ def initialise_glorot_weights(shape, name=None):
         if tf.executing_eagerly():
             return tf.contrib.eager.Variable(initial, name=name)
         return tf.Variable(initial_value=initial, name=name)
+
+
+def normalise(features, normalise_op=tf.nn.l2_normalize):
+    normalised = normalise_op(features, axis=-1)
+    tf.summary.histogram('normalised', normalised)
+    return normalised
+
+
+def chain_aggregate_combine(neighbourhoods, aggregators, combiners, normalisers, name=None):
+    # zip(range(len(r)-1, -1, -1), reversed(r))  # To iterate in reverse with an index. Doesn't play well with zip()
+
+    with tf.name_scope(name, default_name="chain_aggregate_combine") as scope:
+        for i, (aggregator, combiner, normaliser) in enumerate(zip(aggregators, combiners, normalisers)):
+            if i == 0:
+                neighbour_representations = neighbourhoods[i]
+            else:
+                neighbour_representations = full_representations
+
+            neighbourhood_representations = aggregator(neighbour_representations)
+
+            targets = neighbourhoods[i + 1]
+            full_representations = normaliser(combiner(targets, neighbourhood_representations))
+
+        return full_representations
